restructured data | group free time calculations added
