+++ views/userProfile.handlebars
@@ -51,6 +51,10 @@ and their tasks/events alr set up
             <input name="startDate" id="startDate" type="date" required>
             <label for="endDate">End Date:</label>
             <input name="endDate" id="endDate" type="date" required>
+            <label for="startTime">Start Time:</label>
+            <input name="startTime" id="startTime" type="time" required>
+            <label for="endTime">Start Time:</label>
+            <input name="endTime" id="endTime" type="time" required>
             <label for="eventDescription">Event Description:</label>
             <textarea name="eventDescription" id="eventDescription"></textarea>
             <button name="submitButton" id="submitButton" type="submit" value="event">Add Event</button>
@@ -69,6 +73,10 @@ and their tasks/events alr set up
             <input name="startDate" id="startDate" type="date" required>
             <label for="endDate">End Date:</label>
             <input name="endDate" id="endDate" type="date" required>
+            <label for="startTime">Start Time:</label>
+            <input name="startTime" id="startTime" type="time" required>
+            <label for="endTime">Start Time:</label>
+            <input name="endTime" id="endTime" type="time" required>
             <label for="urgencyLevel">Urgency Level:</label>
             <select name="urgencyLevel" id="urgencyLevel">
                 <option value="1">1 (Low)</option>
